modifications de Terms vers badges, size et tagsdatas pour partage vers photoform
