+++ client/src/components/lexiques/TermsTag.jsx
@@ -1,96 +1,9 @@
+import tags from "../tagsData";
+
 import "./styles/Tag.css";
 
 function TermsTag() {
-  const tags = [
-    {
-      title: "Flop",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/buble.svg",
-      alt: "Flop",
-      explanation:
-        "Le flop s’inscrit dans la continuité du tag : plus grand et plus lisible, c’est son prolongement direct. Destiné à la partie vandale, il présente l’avantage d’être réalisé rapidement. En effet, le throw-up consiste à réaliser de manière fluide et rapide un lettrage aux formes simples, plus ou moins arrondies, qui doit être lisible au premier regard. Il est généralement constitué de deux couleurs, la plus foncée servant à réaliser le contour des lettres.\n" +
-        "\n" +
-        "De même que pour le tag, la pratique et l’entraînement peuvent être longs avant de maitriser cette technique. Ce qui peut sembler simple de prime abord se révèle complexe dans la réalisation pour obtenir un flop de qualité. De nombreux writers sont reconnus pour la qualité (et la quantité) de leur throw-up : Seen, JonOne, Cope2, Utah et Ether, Reso, Tilt, Ahero, Horfe, pour ne citer qu’eux !",
-    },
-    {
-      title: "Block Letter",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/cubic.svg",
-      alt: "Block Letter",
-      explanation:
-        "Le block letter est composé d’énormes lettres, de plusieurs mètres de haut. Généralement réalisé le long des voies ferrées, les bords d’autoroute ou des bâtiments désaffectés, car il est plus long à peindre. Certains vont toutefois réaliser des block letters sur des wagons entiers (whole car), voire des trains entiers (whole train).\n" +
-        "\n" +
-        "De formes simples pour être lisibles de loin, les lettres sont généralement réalisées en deux couleurs, à la peinture acrylique à l’aide de rouleaux ou intégralement à la spray. L’objectif du block letter : faire plus gros !",
-    },
-    {
-      title: "Cartoon",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/figure.svg",
-      alt: "Cartoon",
-      explanation:
-        "Le cartoon est un dessin destiné aux films d’animation ou bandes dessinées. Les lignes et les formes réalisées pour ce genre de réalisations sont simples. L’âge d’or du cartoon se situe entre les années 30 et les années 50 avec les productions de Walt Disney ou de Warner Bros. Un graffiti au style cartoon se caractérise par des lignes simples aux couleurs vives et aux représentations rigolotes. On retrouve dans ce style des œuvres figuratives de végétaux ou animaux personnifiées ou des personnages de la culture cinématographique. \n" +
-        "\n" +
-        "Le style cartoon demande du temps de réalisation, car bien que les traits soient simples, le traitement des couleurs peut demander du temps. ",
-    },
-    {
-      title: "Wildstyle",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/wild.svg",
-      alt: "Wildstyle",
-      explanation:
-        "Le wild style porte bien son nom : style sauvage. Le travail des lettres est poussé à l’extrême, à tel point que l’œuvre est quasiment illisible pour un œil non habitué. Les lettres sont déformées, enchevêtrées, enrichies d’ajouts graphiques (flèches, étoiles, vagues…).\n" +
-        "\n" +
-        "L’ajout d’une “épaisseur“ donnant l’impression d’un relief (3D) apporte de la personnalité et améliore la lisibilité aux lettres.\n" +
-        "\n" +
-        "Le travail des couleurs joue également un rôle important dans ce type de graffiti, qui devient une forme globale, graphiquement esthétique.",
-    },
-    {
-      title: "Tag",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/tag.svg",
-      alt: "Tag",
-      explanation:
-        "Le tag est à la base de tous les autres styles. On le définit comme la signature du graffeur. Celui-ci peut être réalisé rapidement et permet d’écrire un grand nombre de fois son pseudonyme. Il est également apposé pour signer un lettrage ou une fresque. Généralement, le taggeur travaille son style au fil des ans, et sa signature devient identifiable au premier coup d’œil pour qui y est attentif.\n" +
-        "C’est un vrai travail de calligraphie : les lettres de l’alphabet deviennent un terrain d’entraînement pour arriver à une signature unique.\n" +
-        "\n" +
-        "Pour beaucoup de graffeurs, le tag reste un critère d’évaluation du niveau technique et de la dextérité d’un peintre : une belle fresque signée avec un tag mal exécuté trahit un manque de pratique dans ce domaine.",
-    },
-    {
-      title: "Small",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/S.svg",
-      alt: "Small",
-      explanation: "Explication pour Ssmall",
-    },
-    {
-      title: "Medium",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/M.svg",
-      alt: "Medium",
-      explanation: "Explication pour Medium",
-    },
-    {
-      title: "Large",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/L.svg",
-      alt: "Large",
-      explanation: "Explication pour Large",
-    },
-    {
-      title: "XtraLarge",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/XL.svg",
-      alt: "XtraL",
-      explanation: "Explication pour XtraLarge",
-    },
-    {
-      title: "Monumental",
-      image:
-        "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/Mtl.svg",
-      alt: "Monumental",
-      explanation: "Explication pour Monumental",
-    },
-  ];
+
 
   return (
     <>
