+++ client/src/components/sizeData.js
@@ -1,0 +1,40 @@
+const size = [
+    {
+        title: "Small",
+        image:
+            "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/S.svg",
+        alt: "Small",
+        explanation: "Explication pour Ssmall",
+    },
+    {
+        title: "Medium",
+        image:
+            "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/M.svg",
+        alt: "Medium",
+        explanation: "Explication pour Medium",
+    },
+    {
+        title: "Large",
+        image:
+            "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/L.svg",
+        alt: "Large",
+        explanation: "Explication pour Large",
+    },
+    {
+        title: "XtraLarge",
+        image:
+            "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/XL.svg",
+        alt: "XtraL",
+        explanation: "Explication pour XtraLarge",
+    },
+    {
+        title: "Monumental",
+        image:
+            "https://raw.githubusercontent.com/WildCodeSchool-2024-02/JS-Nantes-RamCSS-P3-grafitiJam/main/server/public/assets/tags/Mtl.svg",
+        alt: "Monumental",
+        explanation: "Explication pour Monumental",
+    }
+
+];
+
+export default size;
